+++ HashTableArrays/HashTableArrays.iml
@@ -7,6 +7,5 @@
     </content>
     <orderEntry type="inheritedJdk" />
     <orderEntry type="sourceFolder" forTests="false" />
-    <orderEntry type="module" module-name="HashTableArrays" />
   </component>
 </module>
